Moved Dropdown and Leaderboard out of the Leaderboard folder due to CameraCapture using Dropdown
